refactor to formats directory
